+++ lazyscribe/experiment.py
@@ -69,7 +69,7 @@ class Experiment:
     short_slug : str, optional (default None)
         Slugified ``name``. Defaults to calling :py:meth:`slugify.slugify` on the ``name`` attribute.
     slug : str, optional (default None)
-        Unique identifier for the experiment. Deafults to the slugified ``name`` with the creation date
+        Unique identifier for the experiment. Defaults to the slugified ``name`` with the creation date
         appended in the format ``YYYYMMDDHHMMSS``.
     tags : list[str], optional (default [])
         Tags for filtering and identifying experiments across a project.
@@ -388,55 +388,6 @@ class Experiment:
                 fields(Experiment).dirty,
             ),
         )
-
-    def to_tabular(self) -> dict[tuple[str] | tuple[str, str], Any]:
-        """Create a dictionary that can be fed into ``pandas``.
-
-        Returns
-        -------
-        dict
-            Represent the experiment, with the following keys:
-
-            +--------------------------+-------------------------------+
-            | Field                    | Description                   |
-            |                          |                               |
-            +==========================+===============================+
-            | ``("name",)``            | Name of the experiment        |
-            +--------------------------+-------------------------------+
-            | ``("short_slug",)``      | Short slug for the experiment |
-            +--------------------------+-------------------------------+
-            | ``("slug",)``            | Full slug for the experiment  |
-            +--------------------------+-------------------------------+
-            | ``("author",)``          | Experiment author             |
-            +--------------------------+-------------------------------+
-            | ``("last_updated_by",)`` | Last author                   |
-            +--------------------------+-------------------------------+
-            | ``("created_at",)``      | Created timestamp             |
-            +--------------------------+-------------------------------+
-            | ``("last_updated",)``    | Last update timestammp        |
-            +--------------------------+-------------------------------+
-
-            as well as one key per parameter in the ``parameters`` dictionary
-            (with the format ``("parameters", <parameter_name>)``) and one key
-            per metric in the ``metrics`` dictionary (with the format
-            ``("metrics", <metric_name>)``) for each experiment.
-        """
-        d = self.to_dict()
-        return {
-            ("name", ""): d["name"],
-            ("slug", ""): d["slug"],
-            ("short_slug", ""): d["short_slug"],
-            ("author", ""): d["author"],
-            ("created_at", ""): d["created_at"],
-            ("last_updated", ""): d["last_updated"],
-            ("last_updated_by", ""): d["last_updated_by"],
-            **{
-                ("parameters", key): value
-                for key, value in d["parameters"].items()
-                if not isinstance(value, tuple | list | dict)
-            },
-            **{("metrics", key): value for key, value in d["metrics"].items()},
-        }
 
     def promote_artifact(self, repository: Repository, name: str) -> None:
         """Associate an artifact with a :py:class:`lazyscribe.repository.Repository`.