+++ lazyscribe/test.py
@@ -82,40 +82,6 @@ class Test:
             value_serializer=serializer,
         )
 
-    def to_tabular(self) -> dict[tuple[str] | tuple[str, str], Any]:
-        """Create a dictionary that can be fed into ``pandas``.
-
-        Returns
-        -------
-        dict
-            Represent the test, with the following keys:
-
-            +-------------------------------------+-------------------------------+
-            | Field                               | Description                   |
-            |                                     |                               |
-            +=====================================+===============================+
-            | ``("test",)``                       | Test name                     |
-            +-------------------------------------+-------------------------------+
-            | ``("description",)``                | Test description              |
-            +-------------------------------------+-------------------------------+
-
-            as well as one key per parameter in the ``parameters`` dictionary
-            (with the format ``("parameters", <parameter_name>)``) and one key
-            per metric in the ``metrics`` dictionary (with the format
-            ``("metrics", <metric_name>)``) for each test.
-        """
-        d = self.to_dict()
-        return {
-            ("test", ""): d["name"],
-            ("description", ""): d["description"],
-            **{
-                ("parameters", key): value
-                for key, value in d["parameters"].items()
-                if not isinstance(value, tuple | list | dict)
-            },
-            **{("metrics", key): value for key, value in d["metrics"].items()},
-        }
-
 
 @frozen
 class ReadOnlyTest(Test):