+++ tests/test_repository.py
@@ -3,7 +3,6 @@
 import difflib
 import json
 import logging
-import sys
 import warnings
 import zoneinfo
 from datetime import datetime
@@ -82,7 +81,6 @@ def test_save_repository(tmp_path):
             "fname": expected_fname,
             "handler": "json",
             "name": "my-dict",
-            "python_version": ".".join(str(i) for i in sys.version_info[:2]),
             "version": 0,
         },
     ]
@@ -171,7 +169,6 @@ def test_save_repository_multi(tmp_path):
             "fname": "my-dict-20250120132330.json",
             "handler": "json",
             "name": "my-dict",
-            "python_version": ".".join(str(i) for i in sys.version_info[:2]),
             "version": 0,
         },
         {
@@ -179,7 +176,6 @@ def test_save_repository_multi(tmp_path):
             "fname": "my-dict-2-20250120132330.json",
             "handler": "json",
             "name": "my-dict-2",
-            "python_version": ".".join(str(i) for i in sys.version_info[:2]),
             "version": 0,
         },
     ]
@@ -230,7 +226,6 @@ def test_save_repository_multiple_artifact(tmp_path):
             "fname": expected_my_dict_fname0,
             "handler": "json",
             "name": "my-dict",
-            "python_version": ".".join(str(i) for i in sys.version_info[:2]),
             "version": 0,
         },
         {
@@ -238,7 +233,6 @@ def test_save_repository_multiple_artifact(tmp_path):
             "fname": expected_my_dict2_fname,
             "handler": "json",
             "name": "my-dict2",
-            "python_version": ".".join(str(i) for i in sys.version_info[:2]),
             "version": 0,
         },
         {
@@ -246,7 +240,6 @@ def test_save_repository_multiple_artifact(tmp_path):
             "fname": expected_my_dict_fname1,
             "handler": "json",
             "name": "my-dict",
-            "python_version": ".".join(str(i) for i in sys.version_info[:2]),
             "version": 1,
         },
     ]
@@ -518,7 +511,6 @@ def test_retrieve_artifact_meta():
         "fname": "my-dict-20250120132330.json",
         "handler": "json",
         "name": "my-dict",
-        "python_version": ".".join(str(i) for i in sys.version_info[:2]),
         "version": 0,
     }
 
@@ -709,17 +701,3 @@ def test_repository_filter(tmp_path, caplog):
     assert new_spec_["my-metadata"] == repository._search_artifact_versions(
         "my-metadata", 0
     )
-
-
-def test_repository_append_mode_deprecation(tmp_path):
-    """Test reading a repository in append mode."""
-    with warnings.catch_warnings(record=True) as w:
-        warnings.simplefilter("always")
-        repository = Repository(fpath=DATA_DIR / "repository.json", mode="a")
-        assert len(w) == 1
-        assert issubclass(w[-1].category, DeprecationWarning)
-        assert (
-            '`mode="a"` is deprecated and will be removed from `lazyscribe.repository.Repository` in version 2.0'
-            in str(w[-1].message)
-        )
-    assert repository.mode == "w+"